// src/libs/generated/clients/api.schemas.ts
/**
 * Generated by Orval 🍺
 * Do not edit manually.
 * 都道府県別人口構成API
 * RESAS(地域経済分析システム)のデータを元にした、都道府県別の人口構成を取得するAPI。

 * OpenAPI spec version: 1.0.0
 * biome-ignore-all lint: due to automatically generated contents
 */
export interface Prefecture {
  /**
   * 都道府県コード
   * @minimum 1
   * @maximum 47
   */
  prefCode: number;
  /** 都道府県名 */
  prefName: string;
}

export interface PrefecturesResponse {
  /**
   * レスポンスメッセージ
   * @nullable
   */
  message: string | null;
  /** 都道府県のリスト */
  result: Prefecture[];
}

export interface PopulationData {
  /** 年度 */
  year: number;
  /** 人口値 */
  value: number;
  /** 人口割合(%) */
  rate?: number;
}

/**
 * 年齢区分ラベル
 */
export type PopulationCompositionLabel =
  (typeof PopulationCompositionLabel)[keyof typeof PopulationCompositionLabel];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PopulationCompositionLabel = {
  総人口: '総人口',
  年少人口: '年少人口',
  生産年齢人口: '生産年齢人口',
  老年人口: '老年人口',
} as const;

export interface PopulationComposition {
  /** 年齢区分ラベル */
  label: PopulationCompositionLabel;
  /** 年度ごとの人口データ */
  data: PopulationData[];
}

export interface PopulationCompositionResult {
  /** 実績値と推計値の境界年 */
  boundaryYear: number;
  /** 年齢区分別の人口構成データ */
  data: PopulationComposition[];
}

export interface PopulationCompositionPerYearResponse {
  /**
   * レスポンスメッセージ
   * @nullable
   */
  message: string | null;
  result: PopulationCompositionResult;
}

export interface Error {
  /** HTTPステータスコード */
  statusCode?: string;
  /** エラーメッセージ */
  message?: string;
  /** エラーの詳細説明 */
  description?: string;
}

export type GetPopulationCompositionPerYearParams = {
  /**
   * 都道府県コード
   * @minimum 1
   * @maximum 47
   */
  prefCode: number;
};